Fixup %zd. More portable to cast.
git-svn-id: file:///svn/nsd/trunk@2280 a26ef69c-88ff-0310-839f-98b793d9c207

+++ difffile.c
@@ -701,8 +701,8 @@ apply_ixfr(namedb_type* db, FILE *in, const off_t* startpos,
 		ttl = buffer_read_u32(packet);
 		rrlen = buffer_read_u16(packet);
 		if(!buffer_available(packet, rrlen)) {
-			log_msg(LOG_ERR, "bad xfr RR rdata %d, len %d have %zd", 
-				*rr_count, rrlen, buffer_remaining(packet));
+			log_msg(LOG_ERR, "bad xfr RR rdata %d, len %d have %d", 
+				*rr_count, rrlen, (int)buffer_remaining(packet));
 			region_destroy(region);
 			return 0;
 		}
